fixed an issue in the volume (the user should have the permission to read and write to the volume, i was getting an error from mariadb | ERROR 2002 (HY000): Cant connect to local MySQL server through socket /run/mysqld/mysqld.sock(111))

+++ srcs/requirements/wordpress/tools/wordpress-php.sh
@@ -3,25 +3,30 @@
 # added so that we wait for the Mariadb database to be ready
 sleep 10
 # set -e
+until mysql -h mariadb -u"$MADANI_USER" -p"$MADANI_PASSWORD" "$MADANI_DATABASE" -e "SELECT 1" >/dev/null 2>&1; do
+    echo "   MariaDB is unavailable - sleeping"
+    sleep 3
+done
 
-# if [ -f ./wp-config.php ]; then
-# 	echo "WordPress is already installed."
+# Create the folder for the socket is required by php-fpm
+mkdir -p /run/php
 
-# else
+# Give the permissions to www-data user, group to access the folder
+# this is good practice to avoid permission issues, php-fpm comes with its own www-data user and group. 
+# that's why change the ownership of the folder
+chown -R www-data:www-data /run/php
+
+if [ -f /var/www/html/wp-config.php ]; then
+	echo "WordPress is already installed."
+
+else
+	echo "Installing WordPress..."
 # WordPress command line interface (WP-CLI) : provides useful commands and utilities to install, configure, and manage a WordPress site. 
 	curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
 	chmod +x wp-cli.phar
 	mv wp-cli.phar /usr/local/bin/wp-cli
 
 	mkdir -p /var/www/html && cd /var/www/html
-
-	# Create the folder for the socket is required by php-fpm
-	mkdir -p /run/php
-
-	# Give the permissions to www-data user, group to access the folder
-	# this is good practice to avoid permission issues, php-fpm comes with its own www-data user and group. 
-	# that's why change the ownership of the folder
-	chown -R www-data:www-data /run/php
 
 	# Download the WordPress files.
 	wp-cli core download --allow-root
@@ -41,10 +46,15 @@ sleep 10
 						--allow-root
 
 	# Create a new WordPress user.
-	wp-cli user create "$NEW_WP_USER" "$NEW_WP_USER_EMAIL" \
-						--user_pass="$NEW_WP_USER_PASSWORD" \
+	# wp-cli user create "$NEW_WP_USER" "$NEW_WP_USER_EMAIL" \
+	# 					--user_pass="$NEW_WP_USER_PASSWORD" \
+	# 					--role="author" \
+	# 					--allow-root
+	wp-cli user create "mehdi" "mehdi13@gmail.com" \
+						--user_pass="mehdi12345" \
 						--role="author" \
 						--allow-root
-# fi	
+
+fi	
 # finally launch it
 exec /usr/sbin/php-fpm7.4 -F
